requirement.txt
Django==1.10.3
django-taggit==0.21.3
Markdown==2.6.7
pkg-resources==0.0.0
